pyproject.toml
[tool.poetry]
name = "haeserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain = "^0.2.3"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
langchain-chroma = "^0.1.2"
langchainhub = "^0.1.20"
langchain-community = "^0.2.6"
langchain-anthropic = "^0.1.19"
pydantic = "^2.8.2"
supabase = "^2.5.3"
langchain-postgres = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.2.1"}
psycopg2-binary = "^2.9.9"
redis = "^5.0.7"
upstash-redis = "^1.1.0"
langchain-text-splitters = "^0.2.2"
upstash-vector = "^0.5.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-multipart = "^0.0.9"
pyperclip = "^1.9.0"
langchain-pinecone = "^0.1.3"
pinecone-notebooks = "^0.1.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
